requirements.txt
streamlit
instagrapi
pillow